gipsomSyndroma/requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.3
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
jwt==0.6.1
pkg-resources==0.0.0
psycopg2==2.8.3
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.3
